telemetry/requirements.txt
paho-mqtt
motor
pydantic
pydantic-settings
